Capture trace-level logs for OS tests in evergreen
